changes all components to hook
